+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - The formatted duration of a Performance Metric only returns the seconds if duration is below 1 minute
 - The module installer is now called via `bob c:mod` and is prepared to install and uninstall modules (still WIP!)
 - When using `C::Cache()->remember(...)` you can now also specify optional tags to apply to the saved entry
+- Creating a Token via `C::Token()->createToken($length)` is now improved and returns a better string utilizing
+  the whole alphabet instead of base64, also adding `C::Token()->generateSecureToken($length)` for simple token
+  string generation
 
 ---